# apps/api/openapi/api.yaml
openapi: 3.1.0
...
info:
  title: Political Sphere API
  version: 1.0.0
  description: |
    REST API for the Political Sphere platform - a UK-focused multiplayer political simulation.

    This API provides endpoints for:
    - User authentication and management
    - Political party operations
    - Voting and governance actions
    - News and content moderation
    - Game simulation state
    - Real-time multiplayer interactions

    ## Key Features

    - **Democratic Integrity**: All voting operations are cryptographically verifiable
    - **Real-time Updates**: WebSocket integration for live game state
    - **Content Moderation**: AI-assisted moderation with human oversight
    - **Multi-party Support**: Complex coalition and opposition dynamics
    - **Audit Trails**: Complete traceability of all governance actions

    ## Authentication

    The API uses JWT tokens for authentication. Obtain tokens via `/auth/login` or `/auth/register`.

    ## Rate Limiting

    - **Public endpoints**: 100 requests per minute
    - **Authenticated endpoints**: 1000 requests per minute
    - **Voting endpoints**: 10 votes per minute per user
    - **Administrative endpoints**: 100 requests per minute

    ## Versioning

    API versioning follows semantic versioning. Breaking changes increment the major version.
    Current version: v1

    ## Security Considerations

    ### Token Storage
    - Store JWT tokens in httpOnly cookies or secure storage
...

    ### HTTPS Requirements
...

...

...

...

...

...
servers:
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Register new user
      description: Create a new user account with email verification
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
            example:
              email: 'user@example.com'
              password: 'SecurePass123!'
              username: 'political_user'
              firstName: 'John'
              lastName: 'Citizen'
              dateOfBirth: '1990-01-01'
              country: 'GB'
              acceptedTerms: true
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid registration data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - auth
      summary: User login
      description: |
        Authenticate user and return JWT tokens.

        **Rate Limit**: 10 requests per minute per IP address.
        Exceeded limits return HTTP 429 with Retry-After header.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: 'user@example.com'
              password: 'SecurePass123!'
              rememberMe: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout user
      description: Invalidate current session tokens
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Logged out successfully'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{userId}:
    get:
      tags:
        - users
      summary: Get user profile by ID
      description: Retrieve a user's public profile information
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the user
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserProfile'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /parties:
    get:
      tags:
        - parties
      summary: List political parties
      description: Retrieve a paginated list of all political parties
      operationId: listParties
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, memberCount, founded, popularity]
            default: popularity
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Parties retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyListResponse'

    post:
      tags:
        - parties
      summary: Create new political party
      description: Create a new political party (requires authentication)
      operationId: createParty
      security:
        - bearerAuth: []
        - oauth2Auth: [write:parties]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyCreation'
      responses:
        '201':
          description: Party created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '400':
          description: Invalid party data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /votes:
    get:
      tags:
        - votes
      summary: List active votes
      description: Retrieve a list of currently active votes and proposals
      operationId: listActiveVotes
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [legislation, policy, election, referendum, impeachment]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, upcoming, completed]
            default: active
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Active votes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteListResponse'

    post:
      tags:
        - votes
      summary: Create new vote/proposal
      description: Create a new vote or legislative proposal
      operationId: createVote
      security:
        - bearerAuth: []
        - oauth2Auth: [write:votes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteCreation'
      responses:
        '201':
          description: Vote created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '400':
          description: Invalid vote data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Insufficient permissions to create votes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: 'VALIDATION_ERROR'
            message:
              type: string
              description: Human-readable error message
              example: 'The provided data is invalid'
            details:
              type: object
              description: Additional error context
              example:
                field: 'email'
                reason: 'must be a valid email address'
            traceId:
              type: string
              description: Request trace ID for debugging
              example: 'abc-123-def-456'
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
    ValidationError:
      type: object
      required:
        - error
        - validationErrors
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
        validationErrors:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                example: 'email'
              message:
                type: string
                example: 'must be a valid email address'
              code:
                type: string
                example: 'INVALID_FORMAT'
...
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        totalPages:
          type: integer
          minimum: 1
          example: 8

    # Authentication Schemas
    UserRegistration:
      type: object
      required:
        - email
        - password
        - username
        - firstName
        - lastName
        - dateOfBirth
        - country
        - acceptedTerms
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: 'john.citizen@example.com'
        password:
          type: string
          format: password
          minLength: 12
          description: Strong password (min 12 chars, mixed case, numbers, symbols)
          example: 'SecurePass123!'
        username:
          type: string
          pattern: '^[a-zA-Z0-9_-]{3,30}$'
          description: Unique username (3-30 chars, alphanumeric, underscore, hyphen)
          example: 'john_citizen'
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          example: 'John'
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          example: 'Citizen'
        dateOfBirth:
          type: string
          format: date
          description: Date of birth (must be 18+ years ago)
          example: '1990-01-15'
        country:
          type: string
          pattern: '^[A-Z]{2}$'
          description: ISO 3166-1 alpha-2 country code
          example: 'GB'
        acceptedTerms:
          type: boolean
          description: Must be true to register
          example: true
        preferredLanguage:
          type: string
          pattern: '^[a-z]{2}(-[A-Z]{2})?$'
          description: ISO 639-1 language code
          example: 'en-GB'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: 'john.citizen@example.com'
        password:
          type: string
          format: password
          example: 'SecurePass123!'
        rememberMe:
          type: boolean
          default: false
          description: Request longer-lived refresh token
    AuthResponse:
      type: object
      required:
        - user
        - tokens
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        tokens:
          type: object
          required:
            - accessToken
            - refreshToken
            - expiresIn
          properties:
            accessToken:
              type: string
              description: JWT access token
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
            refreshToken:
              type: string
              description: JWT refresh token (store securely)
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
            expiresIn:
              type: integer
              description: Access token expiration time in seconds
              example: 3600
            tokenType:
              type: string
              example: 'Bearer'
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Valid refresh token
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
...
    PublicUserProfile:
      type: object
      description: Publicly visible user information (no sensitive data)
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
          format: uri
        bio:
          type: string
        country:
          type: string
        roles:
          type: array
          items:
            type: string
        stats:
          type: object
          properties:
            votesCast:
              type: integer
            proposalsCreated:
              type: integer
            articlesWritten:
              type: integer
...
    Party:
      type: object
      required:
        - id
        - name
        - description
        - founded
        - status
        - memberCount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174001'
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: 'Progressive Alliance'
        shortName:
          type: string
          maxLength: 10
          example: 'PA'
        description:
          type: string
          minLength: 10
          maxLength: 2000
          example: 'A progressive party focused on social justice and environmental protection.'
        manifesto:
          type: string
          description: Full party manifesto text
          example: 'We believe in...'
        logo:
          type: string
          format: uri
          description: Party logo URL
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: Party brand color (hex)
          example: '#FF6B6B'
        founded:
          type: string
          format: date
          example: '2025-01-01'
        status:
          type: string
          enum: [active, suspended, dissolved]
          example: 'active'
        memberCount:
          type: integer
          minimum: 0
          example: 1250
        leader:
          $ref: '#/components/schemas/PublicUserProfile'
        officers:
          type: array
          items:
            $ref: '#/components/schemas/PublicUserProfile'
        policies:
          type: array
          items:
            type: object
            required:
              - category
              - title
              - summary
            properties:
              category:
                type: string
                enum:
                  [
                    economy,
                    healthcare,
                    education,
                    environment,
                    defense,
                    foreign_policy,
                    justice,
                    technology,
                  ]
              title:
                type: string
              summary:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PartyCreation:
      type: object
      required:
        - name
        - description
        - manifesto
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        shortName:
          type: string
          maxLength: 10
        description:
          type: string
          minLength: 10
          maxLength: 2000
        manifesto:
          type: string
          minLength: 100
        logo:
          type: string
          format: uri
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
...
    PartyListResponse:
      type: object
      required:
        - parties
        - meta
      properties:
        parties:
          type: array
          items:
            $ref: '#/components/schemas/Party'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    VoteListResponse:
      type: object
      required:
        - votes
        - meta
      properties:
        votes:
          type: array
          items:
            $ref: '#/components/schemas/Vote'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # News and Content Schemas
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/refresh

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server admin operations

    mutualTLS:
      type: mutualTLS
      description: |
        Mutual TLS authentication for high-security admin operations.

        Requires client certificate authentication where both the client 
        and server authenticate each other using X.509 certificates. This 
        provides the highest level of security for administrative endpoints.

        **Usage:**
        - Certificate must be issued by a trusted CA
        - Subject DN must match authorized admin principals
        - Certificate must not be expired or revoked
        - Client must present certificate during TLS handshake

        **Applied to:**
        - /admin/audit (audit log access)
        - /admin/users/{userId}/status (user status modifications)

        **Certificate Requirements:**
        - Key size: RSA 2048-bit minimum (Ed25519 preferred)
        - Validity period: Maximum 365 days
        - Extended Key Usage: Client Authentication
        - Subject Alternative Name: admin.political-sphere.com

    oauth2Auth:
      type: oauth2
      flows:
...